requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
contourpy==1.2.0
cryptography==42.0.2
cssutils==2.9.0
cycler==0.12.1
dataframe-image==0.2.3
defusedxml==0.8.0rc2
Django==5.0.2
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
fastjsonschema==2.19.1
fonttools==4.48.1
fpdf==1.7.2
frozenlist==1.4.1
html2image==2.0.4.3
idna==3.6
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
kaleido==0.2.1
kiwisolver==1.4.5
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib==3.8.2
mistune==3.0.2
multidict==6.0.5
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
numpy==1.26.4
oauthlib==3.2.2
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
pillow==10.2.0
platformdirs==4.2.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2024.1
pyzmq==25.1.2
referencing==0.33.0
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
soupsieve==2.5
sqlparse==0.4.4
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.1
tzdata==2024.1
urllib3==2.2.0
webencodings==0.5.1
websocket-client==1.7.0
yarl==1.9.4
